+++ src/indicatorsFormulas/measure/conversionTemplates.ts
@@ -40,10 +40,10 @@ export const prepareConversionParams = (value: TWidgetIndicatorConversionValue) 
     !value.endEventName ||
     !value.endCaseCaseIdFormula ||
     !value.endEventNameFormula ||
-    !value.endEventProcessName ||
+    !value.endEventProcessKey ||
     !value.endEventTimeFormula ||
     !value.startEventNameFormula ||
-    !value.startEventProcessName ||
+    !value.startEventProcessKey ||
     !value.startEventTimeFormula
   ) {
     return null;