+++ src/indicatorsFormulas/shared/createAggregationTemplate.ts
@@ -1,0 +1,10 @@
+function createAggregationTemplate(
+  functionName: string,
+  options?: {
+    distinct?: boolean;
+  }
+) {
+  return `process(${functionName}(${options?.distinct ? "distinct " : ""}{columnFormula}, {eventNameFormula} = '{eventName}'{filters}), {caseCaseIdFormula})`;
+}
+
+export { createAggregationTemplate };