+++ src/widgetContext.ts
@@ -17,9 +17,6 @@ export interface IWidgetTableColumn {
 }
 
 interface IScriptField {
-  // todo: удалить после окончания поддержки миграций [BI-13650]
-  /** @deprecated */
-  guid: string;
   name: string;
   isRequired: boolean;
   isArray: boolean;
@@ -27,9 +24,6 @@ interface IScriptField {
 
 export interface IActionScript {
   key: string;
-  // todo: удалить после окончания поддержки миграций [BI-13650]
-  /** @deprecated */
-  guid: string;
   name: string;
   fields: IScriptField[];
 }
@@ -52,10 +46,6 @@ export interface IGlobalContext {
   reportName: string;
   /** Имена образов по их ключу(в текущем отчете) */
   viewNameByKey: Map<string, string>;
-
-  // todo: удалить после окончания поддержки миграций [BI-13650]
-  /** @deprecated имя группы пространства по ее id */
-  workspaceGroupNameById: Map<number, string>;
 
   /** Меры уровня отчета */
   reportMeasures: TNullable<Map<string, ICommonMeasures>>;
@@ -81,8 +71,10 @@ export interface IGlobalContext {
 
   /** Состояния(название сущности) отчета */
   states: Map<string, ICommonState>;
-  /** Процессы из модели данных */
+  /** @deprecated Процессы из модели данных (по имени) */
   processes: Map<string, IWidgetProcess>;
+  /** Процессы из модели данных (по ключу) */
+  processByKey: Map<string, IWidgetProcess>;
   /** Имена таблиц из модели данных */
   tables: Set<string>;
   /** Функция для запроса информации о колонках таблицы из модели данных */